fix: move values to constants
