+++ packages/scripts/src/utils/contracts.ts
@@ -3,10 +3,6 @@ const hre = require("hardhat");
 import { Contract } from "ethers";
 import { Provider } from "@ethersproject/abstract-provider";
 
-export const FOUNDATION_WALLET = "0x8180d59b7175d4064bdfa8138a58e9babffda44a";
-export const SECONDS_PER_DAY = 86400;
-export const YES_VOTE = "1";
-
 export const getContractInstance = async <T>(
   contractName: DeploymentName | ContractName,
   address?: string,
